backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.0
pydantic-settings==2.5.0
bcrypt==4.0.1
